+++ apps/ledger-live-desktop/src/live-common-set-supported-currencies.js
@@ -6,13 +6,13 @@ setPlatformVersion("0.0.1");
 
 setSupportedCurrencies([
   "bitcoin",
-  "ethereum",
-  "bsc",
+  //"ethereum",
+  //"bsc",
   "polkadot",
   "solana",
   "ripple",
   "litecoin",
-  "polygon",
+  //"polygon",
   "bitcoin_cash",
   "stellar",
   "dogecoin",
@@ -23,7 +23,7 @@ setSupportedCurrencies([
   "tron",
   "tezos",
   "elrond",
-  "ethereum_classic",
+  //"ethereum_classic",
   "zcash",
   "decred",
   "digibyte",
@@ -38,8 +38,8 @@ setSupportedCurrencies([
   "viacoin",
   "stakenet",
   "bitcoin_testnet",
-  "ethereum_ropsten",
-  "ethereum_goerli",
+  //"ethereum_ropsten",
+  //"ethereum_goerli",
   "cosmos_testnet",
   "hedera",
   "cardano",
