--- libs/ledger-live-common/src/families/ethereum/datasets/ethereum1.ts
@@ -1,20 +1,0 @@
-import type { AccountRaw } from "../../../types";
-
-export const ethereum1: AccountRaw = {
-  id: "js:1:ethereum:0x0E3F0bb9516F01f2C34c25E0957518b8aC9414c5:",
-  seedIdentifier: "0x0E3F0bb9516F01f2C34c25E0957518b8aC9414c5",
-  name: "Ethereum legacy xpub6Bem...JyAdpYZy",
-  derivationMode: "",
-  index: 0,
-  freshAddress: "0x0E3F0bb9516F01f2C34c25E0957518b8aC9414c5",
-  freshAddressPath: "44'/60'/0'/0/0",
-  freshAddresses: [],
-  pendingOperations: [],
-  operations: [],
-  currencyId: "ethereum",
-  unitMagnitude: 18,
-  balance: "",
-  blockHeight: 0,
-  lastSyncDate: "",
-  xpub: "",
-};
